Auth - delete pet

+++ src/actions/delete-pet.ts
@@ -2,10 +2,18 @@
 
 import prisma from '@/lib/db';
 import { revalidatePath } from 'next/cache';
-import type { PetID } from '@/lib/types';
 import { petIdSchema } from '@/lib/validations';
+import { auth } from '@/lib/auth';
+import { paths } from '@/lib/paths';
+import { redirect } from 'next/navigation';
 
 export const deletePet = async (petId: unknown) => {
+  const session = await auth();
+
+  if (!session?.user) {
+    redirect(paths.login.path());
+  }
+
   const validatedPetId = petIdSchema.safeParse(petId);
 
   if (!validatedPetId.success) {
@@ -13,6 +21,15 @@ export const deletePet = async (petId: unknown) => {
       message: 'Invalid pet id.',
     };
   }
+
+  const pet = await prisma.pet.findUnique({
+    where: { id: validatedPetId.data },
+    select: { userId: true },
+  });
+
+  if (!pet) return { message: 'Pet not found' };
+
+  if (pet.userId !== session.user.id) return { message: 'Not authorised' };
 
   try {
     await prisma.pet.delete({
